Auth and Authorization code added

+++ FastAPI/fastapi/TODO/routers/todos.py
@@ -6,6 +6,7 @@ from typing import Annotated
 from sqlalchemy.orm import Session 
 from starlette import status
 from pydantic import BaseModel,Field
+from .auth import get_current_user
 
 
 
@@ -24,6 +25,9 @@ def get_db():
 
 db_dependency=Annotated[Session, Depends(get_db)]
 
+# User dependency will make sure authoarization 😀
+user_dependency=Annotated[dict,Depends(get_current_user)]
+
 #request 
 class TodoRequest(BaseModel):
     title:str=Field(min_length=3)
@@ -32,33 +36,42 @@ class TodoRequest(BaseModel):
     complete:bool
 
 
+
+
+@router.post("/todo",status_code=status.HTTP_201_CREATED)
+async def create_todo(user:user_dependency,db:db_dependency,todo_request:TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=401,detail="Authentication Failed")
+    
+    todo_model=Todos(**todo_request.dict(),owner_id=user.get('id'))
+    db.add(todo_model)
+    db.commit()
+
+
 @router.get('/',status_code=status.HTTP_200_OK)
-async def read_all(db: db_dependency ):
-    return db.query(Todos).all()  
+async def read_all(user:user_dependency,db: db_dependency ):
+    if user is None:
+        raise HTTPException(status_code=401,detail="Authentication Failed")
+    return db.query(Todos).filter(Todos.owner_id==user.get("id")).all()  
 
 @router.get('/todo/{todo_id}',status_code=status.HTTP_200_OK)
-async def read_all(db: db_dependency ,todo_id:int=Path(gt=0)):
-    todo_model=db.query(Todos).filter(Todos.id==todo_id).first()
+async def read_by_id(user:user_dependency,db: db_dependency ,todo_id:int=Path(gt=0)):
+    todo_model=db.query(Todos).filter(Todos.owner_id==user.get("id")).filter(Todos.id==todo_id).first()
     if todo_model is not None:
         return todo_model         
     raise HTTPException(status_code=404,detail="Todo not Found")
 
 
-@router.post("/todo",status_code=status.HTTP_201_CREATED)
-async def create_todo(db:db_dependency,todo_request:TodoRequest):
-    todo_model=Todos(**todo_request.dict())
-    db.add(todo_model)
-    db.commit()
-
-
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo( db: db_dependency,
+async def update_todo(user: user_dependency, db: db_dependency,
                       todo_request: TodoRequest,
                       todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401,detail="Authentication Failed")
     
 
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    todo_model = db.query(Todos).filter(Todos.owner_id==user.get("id")).filter(Todos.id == todo_id).first()
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Todo not found.')
 
@@ -71,13 +84,16 @@ async def update_todo( db: db_dependency,
     db.commit()
 
 
-@router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo( db: db_dependency, todo_id: int = Path(gt=0)):
-    # if user is None:
-    #     raise HTTPException(status_code=401, detail='Authentication Failed')
 
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
-        # .filter(Todos.owner_id == user.get('id')).first()
+
+
+@router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
+async def delete_todo( user: user_dependency,db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail='Authentication Failed')
+
+    todo_model = db.query(Todos).filter(Todos.owner_id == user.get('id')).filter(Todos.id == todo_id).first()
+        
     if todo_model is None:
         raise HTTPException(status_code=404, detail='Todo not found.')
     db.query(Todos).filter(Todos.id == todo_id).delete()
